Remove warning message regarding non-zero variance
Part of the approach I took assumed that the first column was the
outcome variable (e.g. dx). I modified the preceding code to insure
that the first column was always the outcome variable. To test this, I
added a modified version of the test data where the dx column was last.