Revert "Switch dontrun to donttest"
This reverts commit 7b091e8182b8ebfc126cf0db48a71f9e6df4bd17.

Because local check() will still run \donttest{} examples
and that takes a while.
CRAN requested we stich \dontrun{} to \donttest{} examples
during manual inspection of the code for the initial submission
to CRAN. But many packages use \dontrun{} for their long-running
tests instead of \donttest{} for this reason.
CRAN likely will not care about it now.