Delete large test fixtures
They were too big for CRAN.
The overall package tarball max size is 5 MB.
Created the tarball with `devtools::build()` and
checked the overall size with `ls -lah`.
Investigated the sizes of contents in the tarball
with `tar -ztvf`, and these test fixtures were by
far the largest files in the tarball.